requirements.txt
-i https://pypi.org/simple
boto3==1.34.50
botocore==1.34.50; python_version >= '3.8'
certifi==2024.2.2; python_version >= '3.6'
jmespath==1.0.1; python_version >= '3.7'
numpy==1.26.4
pinecone-client==3.1.0
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
s3transfer==0.10.0; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
tqdm==4.66.2; python_version >= '3.7'
typing-extensions==4.10.0; python_version >= '3.8'
urllib3==2.0.7; python_version < '3.12' and python_version >= '3.8'
